refactor: BookInfo.tsx 폴더를 book/preview 하위로 이동
